Rename namespace prefixing 'Eclipxe\' vendor
