+++ psalm.xml
@@ -17,8 +17,8 @@
         <LessSpecificReturnType errorLevel="info" />
           <PropertyNotSetInConstructor>
     <errorLevel type="suppress">
-        <referencedProperty name="XmlResourceRetriever\AbstractBaseRetriever::$downloader" />
-        <referencedProperty name="XmlResourceRetriever\Downloader\PhpDownloader::$context" />
+        <referencedProperty name="Eclipxe\XmlResourceRetriever\AbstractBaseRetriever::$downloader" />
+        <referencedProperty name="Eclipxe\XmlResourceRetriever\Downloader\PhpDownloader::$context" />
     </errorLevel>
   </PropertyNotSetInConstructor>
     </issueHandlers>
